requirements.txt
# Auto-generated by generate_requirements.py
# Usage: pip install -r requirements.txt

# base
opencv-python
pillow
numpy
pyfiglet

# ML - TensorFlow
tensorflow

# FastAPI
python-multipart
bcrypt
cryptography
fastapi
pydantic
PyJWT
python-jose
python-multipart
starlette
uvicorn
passlib

# other
scikit-learn
tf-keras

